Counter now glow with the correct "spellID"
Previously they used the glowID for spellID.
It was working, but for the wrong reasons.
Plus, glowing options were sometimes not applied as they should.

It also fixes an issue with spellID in overlays.
It was, again, working for the wrong reasons.
The only counter overlay was Rogue's Riposte which has a single spellID.
The spell uniqueness was the reason why it was working.
But any new "countered overlay" with 2+ spellID would have failed.